+++ frontend/lost/src/components/SIA/lost-sia/src/utils/windowViewport.js
@@ -1,4 +1,4 @@
-export function getViewportCoordinates(wX, wY, svg){
+export function getViewportCoordinates(w, svg){
     const window = {
         xMin:-1*svg.translateX,
         xMax: -1*svg.translateX + (svg.width)/svg.scale,
@@ -15,9 +15,23 @@ export function getViewportCoordinates(wX, wY, svg){
     const scaleX = (viewport.xMax - viewport.xMin)/ (window.xMax - window.xMin)
     const scaleY = (viewport.yMax - viewport.yMin)/ (window.yMax - window.yMin)
 
-    const vX = viewport.xMin + (wX - window.xMin) * scaleX
-    const vY = viewport.yMin + (wY - window.yMin) * scaleY
+    const vX = viewport.xMin + (w.x - window.xMin) * scaleX
+    const vY = viewport.yMin + (w.y - window.yMin) * scaleY
     // console.log('window_viewport: window, viewport', window, viewport)
 
     return {window, viewport, vX, vY, scaleX, scaleY}
+}
+
+/**
+ * 
+ * @param {*} w0 Point in image coordinate system
+ * @param {*} svg Svg with old translation values and old scales
+ * @param {*} s1 New scale/zoom
+ */
+export function getZoomTranslation(w0, svg, s1){
+    const s0 = svg.scale
+    let translation = {x:0,y:0}
+    translation.x = (s0/s1) * (w0.x + svg.translateX) - w0.x 
+    translation.y = (s0/s1) * (w0.y + svg.translateY) - w0.y 
+    return translation
 }