+++ frontend/lost/src/components/SIA/lost-sia/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Always reset annotation mode to *view* when getAnnoBackendFormat is called
 - Delete annotation in sia canvas correctly, when they are moved out of the image.
+- Prevent user from moving image out of canvas
 
 ## [0.0.2] - 2020-04-02
 ### Fixed