+++ frontend/lost/src/components/SIA/lost-sia/CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Always reset annotation mode to *view* when getAnnoBackendFormat is called
 - Delete annotation in sia canvas correctly, when they are moved out of the image.
 - Prevent user from moving image out of canvas
+
+### Fixed
 - Fixed jumping camera when zooming into the image
+- Fixed LabelInput on wrong position when image was zoomed
 
 ## [0.0.2] - 2020-04-02
 ### Fixed